test/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cssselect==1.2.0
dataclasses-json==0.5.7
extractcontent3==0.0.2
faiss-cpu==1.7.4
fastapi==0.95.1
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httplib2==0.22.0
httpx==0.24.0
idna==3.4
importlib-metadata==6.6.0
langchain==0.0.160
lark==1.1.5
lxml==4.9.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
protobuf==4.22.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.6
pyee==8.2.2
pyparsing==3.0.9
pyppeteer==1.0.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
readability-lxml==0.8.1
regex==2023.5.5
requests==2.30.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.12
sse-starlette==1.5.0
sseclient==0.0.27
starlette==0.26.1
tenacity==8.2.2
tiktoken==0.3.3
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.22.0
websockets==10.4
yarl==1.9.2
zipp==3.15.0
